+++ src/Router/Handlers/FallbackExceptionHandler.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Gacela\Router\Handlers;
 
-use Exception;
+use Throwable;
 
 final class FallbackExceptionHandler
 {
-    public function __invoke(Exception $exception): string
+    public function __invoke(Throwable $throwable): string
     {
         header('HTTP/1.1 500 Internal Server Error');