+++ tests/Feature/Router/Fixtures/FakeControllerWithError.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Router\Fixtures;
+
+use TypeError;
+
+final class FakeControllerWithError
+{
+    /**
+     * @throws TypeError
+     */
+    public function __invoke(): string
+    {
+        throw new TypeError('failed');
+    }
+}